requirements.txt
streamlit
langchain
PyMuPDF
python-dotenv
google-generativeai